refactor(automix): move DJ engine into playback/automix.rs
The file named for the concept was a one-line stub; the implementation
lived in player.rs across 17 functions, 5 types, and 3 constants (the
four flag setters, ensure_automix_queue_depth, build_automix_extension*,
order_automix_candidates, decluster_by_album, automix_score and signal
types, weighted_session_shuffle, the metadata fallback, and the helpers
load_queued_external_identities + normalize_external_pair). Move them
into playback/automix.rs so the module's interface, implementation, and
its 19 dedicated tests share one home.

routes.rs now goes through automix::set_automix_*, automix::ensure_automix_queue_depth,
and automix::AUTOMIX_MIN_UPCOMING instead of reaching past the player
module. playback_anchor_index and normalize_genre_key stay in player.rs
but become pub(super) so automix can call them from its sibling slot.
build_session_taste_profile also stays because services/radio.rs depends
on it.

The deletion test passes: removing playback/automix.rs would scatter the
queue-depth orchestrator, scoring, shuffle integration, and "Why" string
plumbing back across five files. The module earns its keep.

All 19 automix tests pass.
